refactor: コードとアーキテクチャのリファクタリング
- Service分離: attendance.service.tsをFacadeパターンと専門サービスに分割
  - ShiftRequestService: シフト申請管理
  - ShiftScheduleService: スケジュール・確定シフト管理
  - AttendanceTrackingService: 出勤記録・打刻管理
  - AttendanceReportingService: ダッシュボード・レポート
  - AttendanceFacadeService: 統合インターフェース

- Server Components最適化: customers/page.tsxをリファクタリング
  - データ取得をサーバーサイドに移行
  - インタラクティブ要素をCustomersPageClientに分離
  - 初期読み込み性能を向上

- 型定義の統合: Zodスキーマを単一の真実の源として使用
  - customer.types.tsとattendance.types.tsをZodスキーマベースに変更
  - 重複する型定義を削除し、z.inferで型を生成
  - 後方互換性のためのエイリアス型を追加
